+++ src/Silverback.Integration.Kafka.Testing/Messaging/Broker/Kafka/Mocks/IInMemoryTopicCollection.cs
@@ -2,6 +2,7 @@
 // This code is licensed under MIT license (see LICENSE file for details)
 
 using System.Collections.Generic;
+using Confluent.Kafka;
 
 namespace Silverback.Messaging.Broker.Kafka.Mocks
 {
@@ -16,6 +17,26 @@ namespace Silverback.Messaging.Broker.Kafka.Mocks
         /// <param name="name">
         ///     The name of the topic.
         /// </param>
-        IInMemoryTopic this[string name] { get; }
+        /// <param name="clientConfig">
+        ///     The client configuration.
+        /// </param>
+        /// <returns>
+        ///     The in-memory topic.
+        /// </returns>
+        IInMemoryTopic Get(string name, ClientConfig clientConfig);
+
+        /// <summary>
+        ///     Gets the topic with the specified name or creates it on the fly.
+        /// </summary>
+        /// <param name="name">
+        ///     The name of the topic.
+        /// </param>
+        /// <param name="bootstrapServers">
+        ///     The bootstrap servers string used to identify the target broker.
+        /// </param>
+        /// <returns>
+        ///     The in-memory topic.
+        /// </returns>
+        IInMemoryTopic Get(string name, string bootstrapServers);
     }
 }